Fixed territories input

+++ app/src/main/java/com/example/riskgame/Risk/views/RiskMapView.java
@@ -45,8 +45,8 @@ public class RiskMapView extends FlashSurfaceView {
     private int top = getTop();
     private int right = getRight();
     private int bottom = getBottom();
-    private float shiftX = 0;
-    private float shiftY = 0;
+    private int shiftX = 0;
+    private int shiftY = 0;
 
     /**
      * MapView
@@ -141,6 +141,7 @@ public class RiskMapView extends FlashSurfaceView {
      */
     public void updatePosition(float changeX, float changeY) {
 
+        //add if checks to prevent scrolling too far
         // updates position variables
         shiftX += changeX;
         shiftY += changeY;
